+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix Uppy init on drag'n drop
 - Add color set option to color field type
 - Fix GraphQL query error requesting select type fields with multiple option
+- Add missing HTTP status codes
 
 ## 2.8.4 (2024-05-05)